[IMP] abstract logic from controller to model
