+++ run.py
@@ -1,8 +1,8 @@
-from app.modules.sleep_spell.models.bot import Bot
+
 from app.config.config import Config
-import subprocess
+from app.modules.sleep_spell.events.sleep_spell import Bot
+
 
 # subprocess.run(["pip", "install", "--upgrade", "pip"], capture_output=True)
 
-
-Bot.start_bot(Config().DISCORD_TOKEN)
+Bot().start_bot(Config().DISCORD_TOKEN)
